app/requirements/dev.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.17.1
Markdown==3.10.2
pillow==12.2.0
sqlparse==0.5.5
typing_extensions==4.15.0